+++ CHANGELOG.md
@@ -1,3 +1,39 @@
+## 0.6.0 (2025-08-25)
+
+### Feat
+
+- incorpora menú interctivo
+
+## 0.5.0 (2025-08-25)
+
+### Feat
+
+- añadir la obtención de información del usuario
+
+## 0.4.1 (2025-08-25)
+
+### Fix
+
+- corregir método de inserción de usuario
+
+## 0.4.0 (2025-08-25)
+
+### Feat
+
+- añadir usuario a tabla hash
+
+## 0.3.0 (2025-08-24)
+
+### Feat
+
+- **archive.py-&-hash.py**: añadir validación de dirección
+
+## 0.2.0 (2025-08-24)
+
+### Feat
+
+- Añadir función hash
+
 ## 0.1.0 (2025-08-24)
 
 ### Feat